# chapter05/project/storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Custom Bikes",
        "contact": {
            "name": "Developer Contact",
            "url": "https://www.example.com"
        },
        "version": 1
    },
    "host": "localhost:8000",
    "basePath": "",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/api/builders": {
            "get": {
                "tags": [
                    "Builders"
                ],
                "summary": "List Builders",
                "responses": {
                    "200": {
                        "description": "Success: List all Builders",
                        "schema": {
                            "$ref": "#/definitions/Builder"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Builders"
                ],
                "summary": "Create Builder",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Json format",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Builder"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Success: A Newly Created Builder",
                        "schema": {
                            "$ref": "#/definitions/Builder"
                        }
                    },
                    "422": {
                        "description": "Missing mandatory field"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "405": {
                        "description": "Invalid HTTP Method"
                    }
                }
            }
        },
        "/api/builders/{id}": {
            "get": {
                "tags": [
                    "Builders"
                ],
                "summary": "Get Builder by Id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Display the specified Builder by id.",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success: Return the Builder",
                        "schema": {
                            "$ref": "#/definitions/Builder"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "405": {
                        "description": "Invalid HTTP Method"
                    }
                }
            },
            "put": {
                "tags": [
                    "Builders"
                ],
                "summary": "Update Builder",
                "description": "Update the specified Builder by Id.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Builder id to update",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Json format",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Builder"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success: Return the Builder updated",
                        "schema": {
                            "$ref": "#/definitions/Builder"
                        }
                    },
                    "422": {
                        "description": "Missing mandatory field"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "405": {
                        "description": "Invalid HTTP Method"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Builders"
                ],
                "summary": "Delete Builder",
                "description": "Delete the specified Builder by id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Builder id to delete",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Success: successful deleted"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "405": {
                        "description": "Invalid HTTP Method"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Builder": {
            "required": [
                "name",
                "description",
                "location"
            ],
            "properties": {
                "name": {
                    "description": "Builder name",
                    "type": "string",
                    "example": "Jesse James"
                },
                "description": {
                    "description": "Famous Motorcycle builder from Texas",
                    "type": "string",
                    "example": "Austin Speed Shop"
                },
                "location": {
                    "description": "Texas/USA",
                    "type": "string",
                    "example": "Austin, Texas"
                }
            }
        },
...
    }
}